# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/pJuridicas": {
            "get": {
                "tags": [
                    "P_Juridica"
                ],
                "summary": "Get a listing of the P_Juridicas.",
                "description": "Get all P_Juridicas",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/P_Juridica"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "P_Juridica"
                ],
                "summary": "Store a newly created P_Juridica in storage",
                "description": "Store P_Juridica",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "P_Juridica that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/P_Juridica"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/P_Juridica"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/pJuridicas/{id}": {
            "get": {
                "tags": [
                    "P_Juridica"
                ],
                "summary": "Display the specified P_Juridica",
                "description": "Get P_Juridica",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of P_Juridica",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/P_Juridica"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "P_Juridica"
                ],
                "summary": "Update the specified P_Juridica in storage",
                "description": "Update P_Juridica",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of P_Juridica",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "P_Juridica that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/P_Juridica"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/P_Juridica"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "P_Juridica"
                ],
                "summary": "Remove the specified P_Juridica from storage",
                "description": "Delete P_Juridica",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of P_Juridica",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/pessoas": {
            "get": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Get a listing of the Pessoas.",
                "description": "Get all Pessoas",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Pessoa"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Store a newly created Pessoa in storage",
                "description": "Store Pessoa",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Pessoa that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Pessoa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Pessoa"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/pessoas/{id}": {
            "get": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Display the specified Pessoa",
                "description": "Get Pessoa",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pessoa",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Pessoa"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Update the specified Pessoa in storage",
                "description": "Update Pessoa",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pessoa",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Pessoa that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Pessoa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Pessoa"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Remove the specified Pessoa from storage",
                "description": "Delete Pessoa",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Pessoa",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "P_Juridica": {
            "required": [
                ""
            ],
            "properties": {
                "cnpj": {
                    "description": "cnpj",
                    "type": "string"
                },
                "nome_fantasia": {
                    "description": "nome_fantasia",
                    "type": "string"
                },
                "inscricao_estadual": {
                    "description": "inscricao_estadual",
                    "type": "string"
                },
                "razao_social": {
                    "description": "razao_social",
                    "type": "string"
                },
                "pessoa_id": {
                    "description": "pessoa_id",
                    "type": "integer",
                    "format": "int32"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "Pessoa": {
            "required": [
                "nome",
                "email"
            ],
            "properties": {
                "nome": {
                    "description": "nome",
                    "type": "string"
                },
                "telefone_1": {
                    "description": "telefone_1",
                    "type": "string"
                },
                "telefone_2": {
                    "description": "telefone_2",
                    "type": "string"
                },
                "email": {
                    "description": "email",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}